feat: split repositories file
