+++ src-tauri/src/repositories/mod.rs
@@ -1,0 +1,2 @@
+pub mod settings;
+pub mod tasks;
